Moved theme functionality to a new service
