+++ src/app/app.component.html
@@ -29,7 +29,7 @@
         </div>
         <div
           class="panel panel-2"
-          [innerHTML]="sideColumnRandomTextItems[0]"></div>
+          [innerHTML]="themePanel2RandomText"></div>
       </div>
       <div class="right-frame-top">
         <div class="banner">{{ appTitle }}</div>
@@ -102,7 +102,7 @@
           <div class="panel panel-5"></div>
           <div
             class="panel panel-6"
-            [innerHTML]="sideColumnRandomTextItems[1]"></div>
+            [innerHTML]="themePanel6RandomText"></div>
           <!-- <div class="panel panel-7"></div> -->
           <!-- <div class="panel panel-8"></div> -->
           <div class="panel panel-9"></div>
