+++ src/app/app.component.ts
@@ -16,6 +16,7 @@ import { AuthService } from './core/auth/auth.service';
 import { RefreshSessionDialogComponent } from './shared/components/refresh-session-dialog/refresh-session-dialog.component';
 import { APP_ROUTES } from './shared/constants/app-routing.constants';
 import { DebuggingService } from './shared/services/debugging.service';
+import { GeneralThemeService } from './shared/services/general-theme.service';
 import { RoutingService } from './shared/services/routing.service';
 
 @Component({
@@ -36,10 +37,11 @@ export class AppComponent implements AfterViewInit {
   isLoggedIn = false;
   autoLogoutCountdown = 0;
   currentYear: number;
-  dataCascade: string;
   showScrollTop = false;
-  sideColumnRandomTextItems: string[] = [];
-  maxNumberOfSideColumnRandomTextItems = 5;
+
+  dataCascade: string;
+  themePanel2RandomText: string;
+  themePanel6RandomText: string;
 
   private warningSubscription: Subscription | undefined;
   private expirySubscription: Subscription | undefined;
@@ -53,6 +55,8 @@ export class AppComponent implements AfterViewInit {
     private routingService: RoutingService,
     private debuggingService: DebuggingService,
     private authService: AuthService,
+    private generalThemeService: GeneralThemeService,
+
     private titleService: Title,
 
     private http: HttpClient,
@@ -75,7 +79,11 @@ export class AppComponent implements AfterViewInit {
         : 'Star Trek Online Info Portal') + this.appTitleTestTag,
     );
     this.currentYear = new Date().getFullYear();
-    this.dataCascade = this.createDynamicDataCascade(8, 7, 3, 6);
+    this.dataCascade = this.generalThemeService.createDynamicDataCascade();
+    this.themePanel2RandomText =
+      this.generalThemeService.createDynamicSideColumnText();
+    this.themePanel6RandomText =
+      this.generalThemeService.createDynamicSideColumnText();
   }
 
   ngOnInit() {
@@ -110,8 +118,6 @@ export class AppComponent implements AfterViewInit {
         }
       },
     );
-
-    this.populateSideColumnRandomTextItems();
   }
 
   ngAfterViewInit() {
@@ -136,80 +142,6 @@ export class AppComponent implements AfterViewInit {
 
   scrollToTop() {
     window.scrollTo({ top: 0, behavior: 'smooth' });
-  }
-
-  populateSideColumnRandomTextItems(): void {
-    for (let i = 0; i < this.maxNumberOfSideColumnRandomTextItems; i++) {
-      this.sideColumnRandomTextItems.push(this.createDynamicSideColumnText());
-    }
-  }
-
-  randomCharacter(): string {
-    const characters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ';
-    return characters.charAt(Math.floor(Math.random() * characters.length));
-  }
-
-  generateRandomValue(minChars: number, maxChars: number): string {
-    const numbers = '0123456789';
-    const special = ' -';
-    let value = '';
-
-    const length =
-      Math.floor(Math.random() * (maxChars - minChars + 1)) + minChars;
-    const numLetters = Math.floor(Math.random() * Math.min(3, length + 1)); // 0 to 2 letters, but not more than the length
-    const numNumbers = length - numLetters;
-
-    for (let i = 0; i < numLetters; i++) {
-      value += this.randomCharacter();
-    }
-
-    for (let i = 0; i < numNumbers; i++) {
-      value += numbers.charAt(Math.floor(Math.random() * numbers.length));
-    }
-
-    // Add space or hyphen with a 1 in 20 chance, but not for the first or last character
-    if (length >= 3 && Math.random() < 1 / 20) {
-      const specialIndex = Math.floor(Math.random() * (length - 3)) + 1;
-      value =
-        value.slice(0, specialIndex) +
-        special.charAt(Math.floor(Math.random() * special.length)) +
-        value.slice(specialIndex);
-    }
-
-    // Shuffle the characters in the value to mix letters and numbers
-    value = value
-      .split('')
-      .sort(() => 0.5 - Math.random())
-      .join('');
-
-    return value;
-  }
-
-  createDynamicDataCascade(
-    rows: number,
-    itemsPerRow: number,
-    minChars: number,
-    maxChars: number,
-  ): string {
-    let html = '';
-
-    for (let i = 1; i <= rows; i++) {
-      html += `<div class="row-${i}">`;
-      for (let j = 1; j <= itemsPerRow; j++) {
-        const value = this.generateRandomValue(minChars, maxChars);
-        html += `<div class="dc${j}">${value}</div>`;
-      }
-      html += '</div>';
-    }
-
-    return html;
-  }
-
-  createDynamicSideColumnText(): string {
-    const value1 = this.generateRandomValue(2, 2);
-    const value2 = this.generateRandomValue(6, 6);
-    const html = `<span class="random-lcars-ref">${value1}<span class="hop">-${value2}</span></span>`;
-    return html;
   }
 
   logout(): void {
